requirements.txt
aiosqlite==0.17.0
flake8==4.0.1
iso8601==0.1.16
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pypika-tortoise==0.1.1
pytz==2021.3
tortoise-orm==0.17.8
typing-extensions==4.0.0
